backend/requirements.txt
appdirs==1.4.3
astroid==2.1.0
attrs==18.2.0
black==18.9b0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pyaml==18.11.0
pyasn1==0.4.4
pyexasol==0.5.1
pylint==2.2.2
PyYAML==3.13
rsa==4.0
six==1.12.0
toml==0.10.0
typed-ast==1.1.1
websocket-client==0.54.0
Werkzeug==0.14.1
wrapt==1.10.11
